Clean up basic styles
Styling directly to html elements isn't ideal, because it might
encourage devs to pick elements based on their styling instead of
selecting the appropriate semantic element. It also creates a cascade
nightmare and requires a lot of undoing styles, which is hard to work
with. Finally, I deleted most of the css variables and am just relying
on Tailwinds implementation, which more people will be familiar with.