Add custom linter and prettier
The default linter is not picking up all whole lot, and I think it is a
better experience to have something more robust. We've also updated the
rules for prettier, which is just a subjective preference. It should
also be noted that we are just running the linter in our code editor (VS
Code) as opposed to reporting to the terminal.